Update image-builder-troubleshooting.md
I encountered an issue where I had the "EC2InstanceProfileForImageBuilder" policy attached to my instance profile, but not "AmazonSSMManagedInstanceCore". Thus, I was experiencing the "TimedOut" issue. I was initially confused because the suggested resolution was what I was doing (had the instance in a default VPC and security group with public subnet) and I was still receiving the error. Once I added the "AmazonSSMManagedInstanceCore" policy, the issue was resolved. I figured I'd clarify that in the documentation in case anyone else experienced this.